Add files via upload
Graph database are the same as relational database that represents data in the form of graph. The nodes are the entities and the relations are the joining of the two nodes. We have used Neo4j, a graph database tool to develop the graphs. We use cypher query to develop the graphs in Neo4j. The text file contains query for generating and analyzing graphs.